+++ web/src/components/inputs/NumberInput.tsx
@@ -1,6 +1,5 @@
 import React, { useState, useEffect, useCallback, memo, useRef } from "react";
-import { createPortal } from "react-dom";
-import { useCombo, useKeyPressedStore } from "../../stores/KeyPressedStore";
+import { useCombo } from "../../stores/KeyPressedStore";
 import PropertyLabel from "../node/PropertyLabel";
 import ThemeNodes from "../themes/ThemeNodes";
 import RangeIndicator from "./RangeIndicator";
@@ -8,376 +7,10 @@ import EditableInput from "./EditableInput";
 import { useFocusPan } from "../../hooks/useFocusPan";
 import { useViewport } from "@xyflow/react";
 import { getMousePosition } from "../../utils/MousePosition";
-// Debug mode
-const DEBUG = false;
-
-// Drag-tuning constants
-const DRAG_THRESHOLD = 10; // px before drag counts
-const DRAG_SLOWDOWN_DEAD_ZONE_PX = 10; // no-slow band above/below slider
-const DRAG_SLOWDOWN_RAMP_BASE_PX = 450; // ramp half-height in px
-const DRAG_SLOWDOWN_RAMP_PX = DRAG_SLOWDOWN_RAMP_BASE_PX * 2; // full ramp span
-const MIN_SPEED_FACTOR = 0.1; // min speed (normal)
-const SHIFT_MIN_SPEED_FACTOR = 0.01; // min speed with Shift
-const SHIFT_SLOWDOWN_DIVIDER = 5; // Shift divides speed by this
-const REFERENCE_SLIDER_WIDTH = 180; // px - base slider width for zoom calculations
-
-// Zoom mapping constants (moved from constants.ts)
-const ZOOM_DRAG_MIN_SCALE = 0.5; // Drag scaling at MIN_ZOOM (18%)
-const ZOOM_DRAG_MAX_SCALE = 2.0; // Drag scaling at MAX_ZOOM (400%)
-
-interface InputProps {
-  nodeId: string;
-  name: string;
-  description?: string | null;
-  min?: number;
-  max?: number;
-  value: number;
-  onChange: (event: any, value: number) => void;
-  id: string;
-  size?: "small" | "medium";
-  color?: "primary" | "secondary";
-  className?: string;
-  inputType?: "int" | "float";
-  hideLabel?: boolean;
-  tabIndex?: number;
-  zoomAffectsDragging?: boolean;
-}
-
-interface NumberInputState {
-  isDefault: boolean;
-  localValue: string;
-  originalValue: number;
-  dragStartX: number;
-  decimalPlaces: number;
-  isDragging: boolean;
-  hasExceededDragThreshold: boolean;
-  dragInitialValue: number;
-  currentDragValue: number;
-  lastClientX: number;
-  dragPointerY: number | null;
-  actualSliderWidth: number;
-}
-
-// Utility Functions
-const calculateStep = (
-  min: number,
-  max: number,
-  inputType: "int" | "float"
-): number => {
-  const range = max - min;
-  let baseStep: number;
-
-  if (inputType === "float") {
-    if (range <= 1) baseStep = 0.01;
-    else if (range <= 20) baseStep = 0.5;
-    else if (range > 20 && range <= 100) baseStep = 0.5;
-    else baseStep = Math.pow(6, Math.floor(Math.log10(range)) - 2);
-  } else {
-    if (range <= 20) baseStep = 0.1;
-    else if (range <= 1000) baseStep = 1;
-    else if (range > 1000 && range <= 5000) baseStep = 16;
-    else if (range > 5000 && range <= 10000) baseStep = 32;
-    else if (range > 10000) baseStep = 64;
-    else baseStep = Math.pow(6, Math.floor(Math.log10(range)) - 1);
-  }
-
-  return baseStep;
-};
-
-const calculateDecimalPlaces = (baseStep: number): number => {
-  return Math.max(0, Math.ceil(Math.log10(1 / baseStep)));
-};
-
-const calculateSpeedFactor = (
-  distanceOutside: number,
-  shiftKey: boolean
-): number => {
-  const t = Math.min(distanceOutside / DRAG_SLOWDOWN_RAMP_PX, 1);
-  let speedFactor = Math.pow(0.1, t);
-  speedFactor = Math.max(speedFactor, MIN_SPEED_FACTOR);
-
-  if (shiftKey) {
-    speedFactor = Math.max(
-      speedFactor / SHIFT_SLOWDOWN_DIVIDER,
-      SHIFT_MIN_SPEED_FACTOR
-    );
-  }
-
-  return speedFactor;
-};
-
-const calculateVisualScreenWidth = (
-  zoomEnabled: boolean,
-  zoom: number
-): number => {
-  return zoomEnabled && zoom > 0
-    ? REFERENCE_SLIDER_WIDTH * zoom
-    : REFERENCE_SLIDER_WIDTH;
-};
-
-const applyValueConstraints = (
-  value: number,
-  min: number,
-  max: number,
-  inputType: "int" | "float",
-  decimalPlaces: number
-): number => {
-  let constrainedValue = value;
-
-  if (inputType === "float") {
-    constrainedValue = parseFloat(value.toFixed(decimalPlaces));
-  } else {
-    constrainedValue = Math.round(value);
-  }
-
-  return Math.max(min, Math.min(max, constrainedValue));
-};
-
-// Custom Hooks
-const useValueCalculation = () => {
-  const calculateStepCb = useCallback(calculateStep, []);
-  const calculateDecimalPlacesCb = useCallback(calculateDecimalPlaces, []);
-  return {
-    calculateStep: calculateStepCb,
-    calculateDecimalPlaces: calculateDecimalPlacesCb
-  };
-};
-
-const useDragHandling = (
-  props: InputProps,
-  state: NumberInputState,
-  setState: React.Dispatch<React.SetStateAction<NumberInputState>>,
-  inputIsFocused: boolean,
-  setInputIsFocused: (focused: boolean) => void,
-  containerRef: React.RefObject<HTMLDivElement>,
-  dragStateRef: React.MutableRefObject<NumberInputState>,
-  setSpeedFactorState: React.Dispatch<React.SetStateAction<number>>,
-  zoom: number
-) => {
-  const { shiftKeyPressed } = useKeyPressedStore((state) => ({
-    shiftKeyPressed: state.isKeyPressed("Shift")
-  }));
-  const { calculateStep, calculateDecimalPlaces } = useValueCalculation();
-
-  const handleMouseMove = useCallback(
-    (e: MouseEvent) => {
-      if (!dragStateRef.current.isDragging) return;
-
-      const { dragStartX, currentDragValue, decimalPlaces, lastClientX } =
-        dragStateRef.current;
-
-      const moveX = e.clientX - dragStartX;
-      if (
-        !dragStateRef.current.hasExceededDragThreshold &&
-        Math.abs(moveX) > DRAG_THRESHOLD
-      ) {
-        dragStateRef.current.hasExceededDragThreshold = true;
-        setInputIsFocused(false);
-      }
-
-      // Ignore tiny drags that haven't crossed the threshold yet
-      if (!dragStateRef.current.hasExceededDragThreshold) {
-        dragStateRef.current.lastClientX = e.clientX;
-        return;
-      }
-
-      if (!containerRef.current) return;
-
-      const rect = containerRef.current.getBoundingClientRect();
-
-      const isOverSlider = e.clientY >= rect.top && e.clientY <= rect.bottom;
-      const isWithinDeadZone =
-        e.clientY >= rect.top - DRAG_SLOWDOWN_DEAD_ZONE_PX &&
-        e.clientY <= rect.bottom + DRAG_SLOWDOWN_DEAD_ZONE_PX;
-
-      // 1) Compute speedFactor for this frame --------------------------------
-      let distanceOutside = 0;
-      if (!isWithinDeadZone) {
-        distanceOutside = Math.max(
-          0,
-          e.clientY < rect.top - DRAG_SLOWDOWN_DEAD_ZONE_PX
-            ? rect.top - DRAG_SLOWDOWN_DEAD_ZONE_PX - e.clientY
-            : e.clientY - (rect.bottom + DRAG_SLOWDOWN_DEAD_ZONE_PX)
-        );
-      }
-
-      const speedFactor = calculateSpeedFactor(distanceOutside, e.shiftKey);
-
-      // expose to overlay/debug
-      setSpeedFactorState(speedFactor);
-
-      //---------------------------------------------------------------------
-
-      const baseStep = calculateStep(
-        props.min ?? 0,
-        props.max ?? 4096,
-        props.inputType || "float"
-      );
-
-      // Always use incremental dragging logic to prevent value jumping
-      const deltaX = e.clientX - lastClientX;
-
-      let newValue: number;
-      if (Math.abs(deltaX) < 0.5) {
-        newValue = currentDragValue;
-      } else {
-        // Step 1: Convert pixel movement to visual percentage
-        const { actualSliderWidth } = dragStateRef.current;
-        const visualScreenWidth = calculateVisualScreenWidth(
-          !!props.zoomAffectsDragging,
-          zoom
-        );
-        const visualPercentage = deltaX / visualScreenWidth;
-
-        // Step 2: Convert to raw value change
-        const range = (props.max ?? 4096) - (props.min ?? 0);
-        const rawValueChange = visualPercentage * range;
-
-        // Step 3: Apply modifiers (speedFactor for vertical slowdown + shift)
-        const effectiveSpeedFactor =
-          isOverSlider && !e.shiftKey ? 1.0 : speedFactor;
-        const finalValueChange = rawValueChange * effectiveSpeedFactor;
-
-        // Step 4: Apply to current value
-        newValue = currentDragValue + finalValueChange;
-
-        // Debug logging
-        if (DEBUG) {
-          console.log("Drag Debug:", {
-            deltaX: deltaX.toFixed(2),
-            actualSliderWidth: actualSliderWidth.toFixed(0),
-            zoom: zoom.toFixed(2),
-            referenceWidth: REFERENCE_SLIDER_WIDTH,
-            visualScreenWidth: visualScreenWidth.toFixed(0),
-            visualPercentage: (visualPercentage * 100).toFixed(2) + "%",
-            range,
-            rawValueChange: rawValueChange.toFixed(4),
-            speedFactor: speedFactor.toFixed(3),
-            effectiveSpeedFactor: effectiveSpeedFactor.toFixed(3),
-            finalValueChange: finalValueChange.toFixed(4),
-            oldValue: currentDragValue.toFixed(4),
-            newValue: newValue.toFixed(4),
-            isOverSlider,
-            shiftKey: e.shiftKey
-          });
-        }
-      }
-
-      // Apply decimal places and value constraints
-      const newDecimalPlaces =
-        props.inputType === "float" ? calculateDecimalPlaces(baseStep) : 0;
-      if (newDecimalPlaces !== decimalPlaces) {
-        dragStateRef.current.decimalPlaces = newDecimalPlaces;
-      }
-
-      newValue = applyValueConstraints(
-        newValue,
-        props.min ?? 0,
-        props.max ?? 4096,
-        props.inputType || "float",
-        newDecimalPlaces
-      );
-
-      if (newValue !== currentDragValue) {
-        dragStateRef.current.currentDragValue = newValue;
-        dragStateRef.current.lastClientX = e.clientX; // reset anchoring only when value actually changes
-        props.onChange(null, newValue);
-      }
-    },
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-    [
-      // Dependencies are stable or managed by refs
-      props.min,
-      props.max,
-      props.inputType,
-      props.onChange,
-      props.zoomAffectsDragging,
-      calculateStep,
-      calculateDecimalPlaces,
-      setInputIsFocused,
-      containerRef,
-      setSpeedFactorState,
-      zoom
-    ]
-  );
-
-  const handleMouseUp = useCallback(() => {
-    if (dragStateRef.current.isDragging) {
-      dragStateRef.current.isDragging = false;
-      // sync final value back to react state
-      setState((prev) => ({
-        ...prev,
-        isDragging: false,
-        localValue: String(dragStateRef.current.currentDragValue)
-      }));
-
-      if (!dragStateRef.current.hasExceededDragThreshold) {
-        setInputIsFocused(true);
-      }
-    }
-  }, [setInputIsFocused, dragStateRef, setState]);
-
-  return { handleMouseMove, handleMouseUp };
-};
-
-const DisplayValue: React.FC<{
-  value: number;
-  isFloat: boolean;
-  decimalPlaces: number;
-}> = ({ value, isFloat, decimalPlaces }) => (
-  <div className="value">
-    {typeof value === "number"
-      ? isFloat
-        ? value.toFixed(decimalPlaces)
-        : value
-      : "NaN"}
-  </div>
-);
-
-// Floating Speed Display Component
-const SpeedDisplay: React.FC<{
-  speedFactor: number;
-  zoom: number;
-  mousePosition: { x: number; y: number };
-  isDragging: boolean;
-  sliderWidth: number;
-}> = ({ speedFactor, zoom, mousePosition, isDragging, sliderWidth }) => {
-  if (!isDragging) return null;
-
-  const speedDisplay = (
-    <div
-      style={{
-        position: "fixed",
-        left: mousePosition.x - 25,
-        top: mousePosition.y + 30,
-        backgroundColor: "rgba(0, 0, 0, 0.9)",
-        color: "white",
-        padding: "6px 10px",
-        borderRadius: "6px",
-        fontSize: "12px",
-        pointerEvents: "none",
-        zIndex: 999999,
-        fontFamily: "monospace",
-        whiteSpace: "nowrap",
-        border: "1px solid rgba(255, 255, 255, 0.2)",
-        boxShadow: "0 2px 8px rgba(0, 0, 0, 0.3)"
-      }}
-    >
-      <div>Speed: {(speedFactor * 100).toFixed(1)}%</div>
-      <div>Zoom: {(zoom * 100).toFixed(0)}%</div>
-      <div>WidthScreen: {(sliderWidth * zoom).toFixed(0)}px</div>
-    </div>
-  );
-
-  // Render the display as a portal to the document body to avoid clipping
-  try {
-    return createPortal(speedDisplay, document.body);
-  } catch (error) {
-    // Fallback: render normally if portal fails
-    console.warn("Failed to create portal for SpeedDisplay:", error);
-    return speedDisplay;
-  }
-};
+import { InputProps, NumberInputState } from "./NumberInput.types";
+import { useDragHandling } from "./NumberInput.hooks";
+import DisplayValue from "./DisplayValue";
+import SpeedDisplay from "./SpeedDisplay";
 
 const NumberInput: React.FC<InputProps> = (props) => {
   const [state, setState] = useState<NumberInputState>({
